[#4310] Implement damage modification config in `DamagesConfig`
Adjust how the `traits-list` partial works to take a customizable
field and modify `DamagesConfig` to pass in a formula input rather
than a checkbox input when handling damage modifications.

This also involves some changes to `TraitAdvancement` so that
damage modification can be registered in `CONFIG.DND5E.traits`
without appearing in the advancement list.